фикс имен, фикс индентации, фикс описаний
