+++ student_223024/BUKEP.Student.WebFormsTodo/Deletion.aspx.cs
@@ -7,8 +7,8 @@ namespace BUKEP.Student.WebFormsTodo
 {
     public partial class Deletion : Page
     {
-		readonly ITaskManager _taskManager = new EFDBTaskManager();
-		private Task _taskToDelete;
+        readonly ITaskManager _taskManager = new EfDatabaseTaskManager();
+        private Task _taskToDelete;
 
         protected void Page_Load(object sender, EventArgs e)
         {
@@ -18,7 +18,7 @@ namespace BUKEP.Student.WebFormsTodo
 
         protected void DeleteButton_Click(object sender, EventArgs e)
         {
-			_taskManager.RemoveTask(_taskToDelete);
+            _taskManager.RemoveTask(_taskToDelete);
             Response.Redirect("~/Todo.aspx");
         }
     }
